Remove dependency on in-memory state when processing EOF junk.
In the case of fatelf-remove, the headers were modified in-memory and
then passed to xappend_junk, which would result in the wrong 'junk'
being appended to the FatELF file.

This change modified xappend_junk to read the FatELF header directly
from the input file, rather than relying on the caller to provide
potentially modified in-memory state. This is a precursor to adding
support for handling Haiku resources via the 'junk' path.